do int before clipping

+++ src/offline_particles/kernel_tools.py
@@ -59,7 +59,7 @@ def split_index(idx: float, max_idx: int) -> tuple[int, float]:
         A tuple containing the integer part and the fractional part.
 
     """
-    int_idx = np.floor(idx)
+    int_idx = int(np.floor(idx))
     int_idx = np.clip(int_idx, 0, max_idx)
     frac_idx = idx - int_idx
-    return int(int_idx), frac_idx
+    return int_idx, frac_idx
